+++ communities/noexport.go
@@ -2,13 +2,13 @@ package communities
 type DoNotSend struct {
 	What       string   `yaml:"what"`
 	Peers string   `yaml:"peers"`
-	Community       int64   `yaml:"community"`
+	Community       int   `yaml:"community"`
 }
 
 type SetLocPref struct {
 	Value       int   `yaml:"value"`
 	Destination string   `yaml:"dest"`
-	Community       int64   `yaml:"community"`
+	Community       int   `yaml:"community"`
 }
 
 type NoExport struct {
